Refactor AffiliateList Component for Improved Structure and Maintainability
Related to #1472
