+++ src/app/components/dashboard/index.ts
@@ -1,7 +1,5 @@
 import { NavBar } from "./NavBar";
-import { AffiliatesList } from "./AffiliatesList";
 
 export {
   NavBar,
-  AffiliatesList
 }
